3rdPartyHelper: custom COM interfaces have been implemented for a better compatibility with different Windows versions. Especially for the compatibility with Windows 8. Also a local cache setting for downloaded files has been added.

+++ Tools/Src/3rdPartyHelper/Program.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Net;
 using System.Security.Cryptography;
-using Shell32;
 
 namespace ThirdPartyHelper
 {
@@ -22,7 +22,25 @@ namespace ThirdPartyHelper
 
             try
             {
-                DownloadAndExtract(args[0]);
+                var downloadCaheFolder = args.Length > 1 ? args[1] : null;
+                if (downloadCaheFolder != null)
+                {
+                    try
+                    {
+                        var folder = Path.GetFullPath(downloadCaheFolder);
+                        if (!Directory.Exists(folder))
+                            Directory.CreateDirectory(folder);
+                        downloadCaheFolder = folder;
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.ForegroundColor = ConsoleColor.Yellow;
+                        Console.WriteLine("WARNING. Argument '{0}' is not valid: {1}.", downloadCaheFolder, ex.Message);
+                        Console.ForegroundColor = defaultColor;
+                        downloadCaheFolder = null;
+                    }
+                }
+                DownloadAndExtract(args[0], downloadCaheFolder);
             }
             catch(Exception ex)
             {
@@ -37,8 +55,8 @@ namespace ThirdPartyHelper
 #endif
         }
 
-        static void DownloadAndExtract(string fileName)
-        {
+        static void DownloadAndExtract(string fileName, string caheFolder)
+        {   
             var file = Path.GetFullPath(fileName);
             if (!File.Exists(file))
                 throw new FileNotFoundException("File is not found.", file);
@@ -63,42 +81,71 @@ namespace ThirdPartyHelper
 
             foreach (var item in downloadList)
             {
-                var tempFile = Path.GetTempFileName() + ".zip";
-                HttpGet(item.Uri, tempFile);
-
-
-
-                using (var fStream = new FileStream(tempFile, FileMode.Open, FileAccess.Read))
+                bool downloaded = false, keepFile = false;
+                string tempFile;
+                if (caheFolder != null && !string.IsNullOrEmpty(item.CacheFileName))
                 {
-                    var hash = md5.ComputeHash(fStream);
-                    var hashString = BitConverter.ToString(hash).Replace("-", string.Empty).ToLowerInvariant();
-                    var color = Console.ForegroundColor;
-                    Console.ForegroundColor = ConsoleColor.Yellow;
-                    if (string.IsNullOrEmpty(item.HashCode))
+                    tempFile = Path.GetFullPath(Path.Combine(caheFolder, item.CacheFileName));
+                    var tempDir = Path.GetDirectoryName(tempFile);
+                    Debug.Assert(!string.IsNullOrEmpty(tempDir));
+                    if (!Directory.Exists(tempDir))
+                        Directory.CreateDirectory(tempDir);
+                    else if (File.Exists(tempFile))
                     {
-                        Console.WriteLine("WARNING. Hash code is not defined for the file. Hash code check will be skipped.");
-                        Console.WriteLine("Hash code: {0}", hashString);
+                        using (var fStream=new FileStream(tempFile,FileMode.Open,FileAccess.Read))
+                        {
+                            var hash = md5.ComputeHash(fStream);
+                            var hashString = BitConverter.ToString(hash).Replace("-", string.Empty).ToLowerInvariant();
+                            downloaded = hashString == item.HashCode;
+                            if (downloaded)
+                                Console.WriteLine("The file from {0} has been found in the cache.", item.Uri);
+                        }
                     }
-                    else if (!string.Equals(item.HashCode, hashString, StringComparison.InvariantCultureIgnoreCase))
+                    keepFile = true;
+                }
+                else
+                    tempFile = Path.GetTempFileName() + ".zip";
+
+                if (!downloaded)
+                {
+                    HttpGet(item.Uri, tempFile);
+
+                    using (var fStream = new FileStream(tempFile, FileMode.Open, FileAccess.Read))
                     {
-                        Console.WriteLine("WARNING. Hash code mismatch.");
-                        Console.WriteLine("Hash code: {0}. Required hash code: {1}.", hashString, item.HashCode);
+                        var hash = md5.ComputeHash(fStream);
+                        var hashString = BitConverter.ToString(hash).Replace("-", string.Empty).ToLowerInvariant();
+                        var color = Console.ForegroundColor;
+                        Console.ForegroundColor = ConsoleColor.Yellow;
+                        if (string.IsNullOrEmpty(item.HashCode))
+                        {
+                            Console.WriteLine(
+                                "WARNING. Hash code is not defined for the file. Hash code check will be skipped.");
+                            Console.WriteLine("Hash code: {0}", hashString);
+                        }
+                        else if (!string.Equals(item.HashCode, hashString, StringComparison.InvariantCultureIgnoreCase))
+                        {
+                            Console.WriteLine("WARNING. Hash code mismatch.");
+                            Console.WriteLine("Hash code: {0}. Required hash code: {1}.", hashString, item.HashCode);
+                        }
+                        Console.ForegroundColor = color;
                     }
-                    Console.ForegroundColor = color;
                 }
 
                 UnzipAll(shell, tempFile, item.SourcePath, item.DestinationPath);
 
-                try
+                if (!keepFile)
                 {
-                    File.Delete(tempFile);
-                }
-                catch
-                {
-                    var color = Console.ForegroundColor;
-                    Console.ForegroundColor = ConsoleColor.Yellow;
-                    Console.WriteLine("WARNING. Can't delete temporary file '{0}'.", tempFile);
-                    Console.ForegroundColor = color;
+                    try
+                    {
+                        File.Delete(tempFile);
+                    }
+                    catch
+                    {
+                        var color = Console.ForegroundColor;
+                        Console.ForegroundColor = ConsoleColor.Yellow;
+                        Console.WriteLine("WARNING. Can't delete temporary file '{0}'.", tempFile);
+                        Console.ForegroundColor = color;
+                    }
                 }
             }
 
@@ -130,7 +177,7 @@ namespace ThirdPartyHelper
             }
         }
 
-        static void UnzipAll(ShellClass shell, string zippedFile, string sourcePath, string destinationFolder)
+        private static void UnzipAll(ShellClass shell, string zippedFile, string sourcePath, string destinationFolder)
         {
             var destinationPath = Path.GetFullPath(destinationFolder);
             Console.WriteLine("Extracting files to {0}", destinationPath);
@@ -140,24 +187,10 @@ namespace ThirdPartyHelper
 
             var source = Path.Combine(Path.GetFullPath(zippedFile), sourcePath);
 
-            Folder from = null, to = null;
-            try
-            {
-                from = shell.NameSpace(source);
-                to = shell.NameSpace(destinationPath);
+            var from = shell.NameSpace(source);
+            var to = shell.NameSpace(destinationPath);
 
-                to.CopyHere(from.Items(), 0);
-            }
-            finally
-            {
-                var disp = to as IDisposable;
-                if (disp != null)
-                    disp.Dispose();
-
-                disp = from as IDisposable;
-                if (disp != null)
-                    disp.Dispose();
-            }
+            to.CopyHere(from.Items(), 0);
         }
     }
 }
